feat(topology-vm) : removing css changes
