+++ plugins/topology/src/components/Topology/TopologyComponent.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 
-import { makeStyles } from '@material-ui/core';
 import { useTheme } from '@material-ui/core/styles';
 
 import { FilterContext } from '../../hooks/FilterContext';
@@ -50,19 +49,11 @@ export const TopologyComponent = () => {
 
   const k8sResourcesContextData = useK8sObjectsResponse(watchedResources);
   const filterContextData = useFilterContextValues();
-  // below lines added due to custom styles not working once started working remove usestyles and replace className={classes.pfRITopology} -> className="pf-ri__topology"
-  const useStyles = makeStyles(theme => ({
-    pfRITopology: {
-      height: '100%',
-      display: 'flex',
-      flexDirection: 'column',
-    },
-  }));
-  const classes = useStyles();
+
   return (
     <K8sResourcesContext.Provider value={k8sResourcesContextData}>
       <FilterContext.Provider value={filterContextData}>
-        <div className={classes.pfRITopology}>
+        <div className="pf-ri__topology">
           <TopologyWorkloadView />
         </div>
       </FilterContext.Provider>
