+++ plugins/topology/src/data-transforms/data-transformer.ts
@@ -4,6 +4,7 @@ import { Model, NodeModel, NodeShape } from '@patternfly/react-topology';
 import { TYPE_APPLICATION_GROUP, TYPE_VM, TYPE_WORKLOAD } from '../const';
 import { CronJobModel } from '../models';
 import { K8sResponseData, K8sWorkloadResource } from '../types/types';
+import { VM_TYPES } from '../types/vms';
 import { getPipelinesDataForResource } from '../utils/pipeline-utils';
 import { getPodsDataForResource } from '../utils/pod-resource-utils';
 import {
@@ -17,7 +18,6 @@ import {
 } from '../utils/resource-utils';
 import {
   createTopologyNodeData,
-  VM_TYPES,
   WORKLOAD_TYPES,
 } from '../utils/topology-utils';
 import {
