+++ plugins/topology/src/types/vms.ts
@@ -1,0 +1,3 @@
+import { ModelsPlural } from '../models';
+
+export const VM_TYPES: string[] = [ModelsPlural.virtualmachines];
